Fix DataTables breaking on Events Dashboard
Apparently, initializing DataTables on tables inside elements with
`display: none;` breaks their sizing. This is because DataTable is
unable to compute the size of such elements.

This was fixed by initializing DataTables only after the tab has
been opened. Also fixed updating the URL hash according to which
tab is open.

Also move the JS into a separate file and make the tables full-width.

Fixes #3252