Put get_new_identifier method on Event class
The first thing I've done here is move the method onto the class - it seemed untidy loitering at the top of the file. I updated the method name, so that it's more closely aligned with the method on the Order model, and then updated various places where it's called. This makes everything neater.

I've noticed that this method generates a different format of UUID to the method in Order, but that it's similar to the method SocialLink.get_new_social_link_identifier. I wonder if the two can be combined? If so, I would add the method to the base class with a different name and have both SocialLink and Event call it.