All kinds of mails should be considered.
An email is a string (a subset of ASCII characters) separated into two parts by @ symbol, a personal info and a domain. The personal info part and domain part should consider any sort of them, and should be restricted. I have tried to consider all alphabets, alphanumeric characters etc. separate them by a @ and then consider all kinds of domain. So, that it is not only restricted to some mail ids.